==> S&P-500/02. api-pull.ipynb <==
import numpy as np
import pandas as pd
import requests
import json
import time

df = pd.read_csv('01b. Stock-Background.csv')

df = df[['Symbol', 'Security', 'GICS Sector']]

ticker_list = np.array(df[['Symbol']]).tolist()

print(ticker_list)

# S&P 500 doesn't have any special characters in their stock-list, so no need for cleaning right now.

ticker_list_clean = [i[0] for i in ticker_list]

ticker_list_clean

ticker_list_clean[371:]

def get_ticker_json(ticker_name):
    url = f"https://www.alphavantage.co/query?function=TIME_SERIES_WEEKLY_ADJUSTED&symbol={ticker_name}&apikey=[KEYHERE]"
    r = requests.get(url)
    data = r.json()
    with open("./S&P-data/" + str(ticker_name) + ".json", 'w') as f:
        json.dump(data, f)

get_ticker_json('MMM')

r = requests.get(url)
data = r.json()

with open('./S&P-data/data.json', 'w') as f:
    json.dump(data, f)

for i in ticker_list_clean:
    print(i)
    get_ticker_json(i)
    time.sleep(12.5)

ticker_list_clean[372:]

for i in ticker_list_clean[372:]:
    print(i)
    get_ticker_json(i)
    time.sleep(12.5)

for i in ticker_list_clean[481:]:
    print(i)
    get_ticker_json(i)
    time.sleep(13)
