==> S&P-500/04. Bulk-transform.ipynb <==
import pandas as pd
from pathlib import Path
import json

calendar = pd.read_csv('datetime_index.csv', index_col=1)

calendar.index = calendar.index.rename('Date')
calendar = calendar.rename(columns={'Unnamed: 0': 'Increment'})

calendar

p = Path(r'D:\Documents\Datasets\001. ML Projects\Sticker\S&P-500\S&P-data').glob('*.json')

file_names = []
for i in p:
    file_names.append(i)

def json_to_df(file):
    data_dict = {}
    with open(file, 'r') as f:
        
        data = json.load(f)
        name = data['Meta Data']['2. Symbol']
        for i in data['Weekly Adjusted Time Series']:
            week = str(i)

            close = str(data['Weekly Adjusted Time Series'][i]['5. adjusted close'])
            data_dict[week] = close
        
        stock_df = pd.DataFrame.from_dict(data_dict, orient='index')
        stock_df = stock_df.rename(columns={0: name})
    f.close()
    return stock_df

cal_df = calendar
for i in file_names:
    print(i)
    stock_df = json_to_df(i)
    cal_df = cal_df.join(stock_df)

df_out = cal_df.drop(columns='Increment').dropna(how='all')

df_out.to_csv('S&P_stock_prices.csv')

for i in df_out.columns:
    df_out[i] = pd.to_numeric(df_out[i])

diff_df = df_out.diff()

diff_df.to_csv('S&P_stock_differences.csv')
