==> S&P-500/07. load data.ipynb <==
import numpy as np
import pandas as pd
import json
from pathlib import Path
import boto3
from decimal import Decimal
import time

with open(Path(r'../access.json'), 'r') as access_file:
    access = json.load(access_file)
    access_file.close()

data_path = Path(r'./S&P-sorted-data/').glob('*.json')

meta = Path(r'./meta-data/')

for i in data_path:
    meta_path = meta.joinpath(i.name)
    with open(meta_path, 'r') as meta_file:
        meta_data = json.load(meta_file)
        
        try:
            name = meta_data["Name"]
            print(name)
        except:
            name = None
        
        try:
            description = meta_data["Description"]
        except:
            description = None
        
        try:
            sector = meta_data["Sector"]
        except:
            sector = None

        try:
            industry = meta_data["Industry"]
        except:
            industry = None
        
        try:
            yearhigh = meta_data["52WeekHigh"]
            print(yearhigh)
        except:
            yearhigh = None

        try:
            yearlow = meta_data["52WeekLow"]
            print(yearlow)
        except:
            yearlow = None
        
        try:
            dividend = meta_data["DividendPerShare"]
        except:
            dividend = None



dynamodb = boto3.resource('dynamodb',region_name='us-east-2',aws_access_key_id=access['KEY_IAM_AWS'],aws_secret_access_key=access['SECRET_IAM_AWS'])

table = dynamodb.Table('tickers_sp')

def smaller_dict(d: dict()):
    list = [{k: v} for k, v in reversed(d.items())]
    return list

for i in data_path:
    meta_path = meta.joinpath(i.name)
    row_dict = {}
    with open(i, 'r') as json_file:
            print(i.stem)
            row_dict['ticker'] = i.stem
            
            data = json.load(json_file)
            
            
            row_dict['price'] = smaller_dict(data[i.stem]['price'])

            try:
                row_dict['pos_vals'] = data[i.stem]['pos_vals']
            except:
                row_dict['pos_vals'] = None

                print(f"skipping pos_vals for {i}")

            try:
                row_dict['neg_vals'] = data[i.stem]['neg_vals']
            except:
                row_dict['neg_vals'] = None
                
                print(f"skipping neg_vals for {i}")

            try:
                row_dict['dec_vals'] = data[i.stem]['decorrelated_vals']
            except:
                print(f"skipping neg_vals for {i}")
                row_dict['dec_vals'] = None

    json_file.close()
    
    with open(meta_path, 'r') as meta_file:
        meta_data = json.load(meta_file)
        try:
            row_dict['name'] = meta_data["Name"]
        except:
            name = None
            print(f"skipping name for {i}")

        try:
            row_dict['description'] = meta_data["Description"]
        except:
            row_dict['description'] = None
            print(f"skipping description for {i}")

        try:
            row_dict['industry'] = meta_data['Industry']
        except:
            row_dict['industry'] = None
            print(f"skipping industry for {i}")   

        try:
            row_dict['sector'] = meta_data['Sector']
        except:
            row_dict['sector'] = None
            print(f"skipping sector for {i}") 
        try:
            row_dict['yearhigh'] = meta_data["52WeekHigh"]
        except:
            row_dict['yearhigh'] = None
            print(f"skipping yearhigh for {i}") 

        try:
            row_dict['yearlow'] = meta_data["52WeekLow"]
        except:
            row_dict['yearlow'] = None
            print(f"skipping yearlow for {i}") 

        try:
            row_dict['dividend'] = meta_data["DividendPerShare"]
        except:
            row_dict['dividend'] = None   
            print(f"skipping dividend for {i}")

        meta_file.close()

    table.put_item(Item=json.loads(json.dumps(row_dict), parse_float=Decimal))
    time.sleep(1.5)
